Get UForge version automatically
- improvements
